Added "OK" Button to the disorder detail page (to leave the page), added symptom part
