+++ BFH_USZ_PICC/BFH_USZ_PICC/Controls/EmergencyFlyout.xaml.cs
@@ -20,32 +20,10 @@ namespace BFH_USZ_PICC.Controls
         public EmergencyFlyout()
         {
             InitializeComponent();
-            
+
             //Binds all possible symptoms to the binding context
             BindingContext = new DisorderViewModel();
         }
 
-        void ContactUSZTelemedizin(object sender, EventArgs e)
-        {
-            if(DependencyService.Get<ICaller>().CanMakePhonecall())
-            {
-                CallUSZTelemedizin();
-            }
-            else
-            {
-                App.Current.MainPage.DisplayAlert("Fehlgeschlagen", "Ihr Gerät kann die USZ Telemedizin nicht anrufen.\nKontaktieren sie 044 666 66 66", "Ok");
-            }
-        }
-
-        async void CallUSZTelemedizin()
-        {
-            bool call = await App.Current.MainPage.DisplayAlert("Warnung", "USZ Telemedizin wirklich anrufen?", "Ja", "Nein");
-            if (call)
-            {
-                var dialer = DependencyService.Get<ICaller>();
-                if (dialer != null)
-                    dialer.Dial("0764979662");
-            }
-        }
     }
 }
