+++ BFH_USZ_PICC/BFH_USZ_PICC/Views/DisorderDetailPage.xaml.cs
@@ -1,4 +1,5 @@
-﻿using BFH_USZ_PICC.Models;
+﻿using BFH_USZ_PICC.Interfaces;
+using BFH_USZ_PICC.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -15,7 +16,34 @@ namespace BFH_USZ_PICC.Views
         {
             InitializeComponent();
             BindingContext = selectedEntry;
+        }
 
+        void LeavePage(object sender, EventArgs e)
+        {
+            Navigation.PopModalAsync();
+        }
+
+        void ContactUSZTelemedizin(object sender, EventArgs e)
+        {
+            if (DependencyService.Get<ICaller>().CanMakePhonecall())
+            {
+                CallUSZTelemedizin();
+            }
+            else
+            {
+                App.Current.MainPage.DisplayAlert("Fehlgeschlagen", "Ihr Gerät kann die USZ Telemedizin nicht anrufen.\nKontaktieren sie 044 666 66 66", "Ok");
+            }
+        }
+
+        async void CallUSZTelemedizin()
+        {
+            bool call = await App.Current.MainPage.DisplayAlert("Warnung", "USZ Telemedizin wirklich anrufen?", "Ja", "Nein");
+            if (call)
+            {
+                var dialer = DependencyService.Get<ICaller>();
+                if (dialer != null)
+                    dialer.Dial("0764979662");
+            }
         }
     }
 }
